# servers/fastapi/openapi_spec.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Presenton MCP - Curated Tools",
    "description": "Essential Presenton API endpoints for MCP chatbot integration (10 tools)",
    "version": "1.1.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000",
      "description": "Local Presenton API"
    },
    {
      "url": "http://127.0.0.1:5002",
      "description": "Slide Helper API (token-efficient single slide retrieval)"
    }
  ],
  "paths": {
...
    "/api/v1/ppt/slide/{id}": {
      "get": {
        "operationId": "get_slide",
        "summary": "Get single slide by ID (token-efficient)",
        "description": "Retrieve a specific slide without fetching the entire presentation. More token-efficient than get_presentation for single slide operations. Use this before editing a slide to get the latest content.",
        "servers": [{"url": "http://127.0.0.1:5002"}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Slide UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response with slide data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "404": {
            "description": "Slide not found"
          }
        }
      },
      "delete": {
        "operationId": "delete_slide",
        "summary": "Delete slide by ID",
        "description": "Delete a specific slide from its presentation. This removes the slide from the slides array and re-indexes remaining slides. Use this to permanently remove a slide.",
        "servers": [{"url": "http://127.0.0.1:5002"}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Slide UUID to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "Slide deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "presentation_id": {
                      "type": "string"
                    },
                    "slides_remaining": {
                      "type": "integer"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Slide not found"
          }
        }
      }
    },
...
  }
}